+++ packages/css-color-parser/src/color-data.ts
@@ -544,7 +544,7 @@ export function noneToZeroInRelativeColorDataChannels(x: Map<string, TokenNumber
 
 function dummyNumberToken(x: number): TokenNumber {
 	if (Number.isNaN(x)) {
-		x = 0;
+		return [TokenType.Number, 'none', -1, -1, { value: Number.NaN, type: NumberType.Number }];
 	}
 
 	return [TokenType.Number, x.toString(), -1, -1, { value: x, type: NumberType.Number }];